GetbyId and GetAll Implementation

+++ GeekShopping/GeekShopping.ProductAPI/Repository/IProductRepository.cs
@@ -5,7 +5,7 @@ namespace GeekShopping.ProductAPI.Repository
     public interface IProductRepository
     {
         Task<IEnumerable<ProductVO>> FindAll();
-        Task<ProductVO> FindById(int id);
+        Task<ProductVO> FindById(long id);
         Task<ProductVO> Create(ProductVO vo);
         Task<ProductVO> Update(ProductVO vo);
         Task<bool> Delete(long id);
